ROCm implementation of nvdiffrast

+++ csrc/common/hipraster/impl/CoarseRaster.inl
@@ -151,7 +151,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
                 // First warp: Choose the segment with the lowest initial triangle index.
 
                 bool hasStream = (thrInBlock < CR_BIN_STREAMS_SIZE);
-                U32 hasStreamMask = __ballot_sync(~0u, hasStream);
+                U32 hasStreamMask = __ballot_sync((unsigned long long)(~0u), hasStream);
                 if (hasStream)
                 {
                     // Find the stream with the lowest triangle index.
@@ -180,7 +180,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
                     // Consume and broadcast.
 
                     bool first = (s_scanTemp[0][CR_BIN_STREAMS_SIZE - 1 + 16] == firstTri);
-                    U32 firstMask = __ballot_sync(hasStreamMask, first);
+                    U32 firstMask = __ballot_sync((unsigned long long)(hasStreamMask), first);
                     if (first && (firstMask >> threadIdx.x) == 1u)
                     {
                         int segIdx = s_binStreamCurrSeg[thrInBlock];
@@ -246,7 +246,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
 
             // 32 triangles per warp: Record emits (= tile intersections).
 
-            if (__any_sync(~0u, triIdx != -1))
+            if (__any_sync((unsigned long long)(~0u), triIdx != -1))
             {
                 S32 v0x = sub_s16lo_s16lo(triData.x, originX);
                 S32 v0y = sub_s16hi_s16lo(triData.x, originY);
@@ -273,7 +273,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
 
                 // Case A: All AABBs are small => record the full AABB using atomics.
 
-                if (__all_sync(~0u, sizex <= 2 && sizey <= 2))
+                if (__all_sync((unsigned long long)(~0u), sizex <= 2 && sizey <= 2))
                 {
                     if (triIdx != -1)
                     {
@@ -326,26 +326,26 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
                     d12x += sizex * d12y;
 
                     // Case B: Warp-AABB is not much larger than largest AABB => Check tiles in warp-AABB, record using ballots.
-                    if (__any_sync(~0u, warea * 4 <= area * 8))
+                    if (__any_sync((unsigned long long)(~0u), warea * 4 <= area * 8))
                     {
                         // Not sure if this is any faster than Case C after all the post-Volta ballot mask tracking.
                         bool act = (triIdx != -1);
-                        U32 actMask = __ballot_sync(~0u, act);
+                        U32 actMask = __ballot_sync((unsigned long long)(~0u), act);
                         if (act)
                         {
                             for (int y = wloy; y <= whiy; y++)
                             {
                                 bool yIn = (y >= loy && y <= hiy);
-                                U32 yMask = __ballot_sync(actMask, yIn);
+                                U32 yMask = __ballot_sync((unsigned long long)(actMask), yIn);
                                 if (yIn)
                                 {
                                     for (int x = wlox; x <= whix; x++)
                                     {
                                         bool xyIn = (x >= lox && x <= hix);
-                                        U32 xyMask = __ballot_sync(yMask, xyIn);
+                                        U32 xyMask = __ballot_sync((unsigned long long)(yMask), xyIn);
                                         if (xyIn)
                                         {
-                                            U32 res = __ballot_sync(xyMask, b01 >= 0 && b02 >= 0 && b12 >= 0);
+                                            U32 res = __ballot_sync((unsigned long long)(xyMask), b01 >= 0 && b02 >= 0 && b12 >= 0);
                                             if (threadIdx.x == 31 - __clz(xyMask))
                                                 *(U32*)currPtr = res;
                                             currPtr += 4, b01 -= d01y, b02 += d02y, b12 -= d12y;
@@ -391,7 +391,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
             {
                 int tileInBin = tileInBin_base + thrInBlock;
                 bool act = (tileInBin < CR_BIN_SQR);
-                U32 actMask = __ballot_sync(~0u, act);
+                U32 actMask = __ballot_sync((unsigned long long)(~0u), act);
                 if (act)
                 {
                     // Compute prefix sum of emits over warps.
@@ -415,10 +415,10 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
 
                     // All counters within the warp are small => compute prefix sum using ballot.
 
-                    if (!__any_sync(actMask, tileEmits >= 2))
+                    if (!__any_sync((unsigned long long)(actMask), tileEmits >= 2))
                     {
                         U32 m = getLaneMaskLe();
-                        *v = (__popc(__ballot_sync(actMask, tileEmits & 1) & m) << emitShift) | __popc(__ballot_sync(actMask, tileAllocs & 1) & m);
+                        *v = (__popc(__ballot_sync((unsigned long long)(actMask), tileEmits & 1) & m) << emitShift) | __popc(__ballot_sync((unsigned long long)(actMask), tileAllocs & 1) & m);
                     }
 
                     // Otherwise => scan-32 within the warp.
@@ -441,7 +441,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
             __syncthreads();
 
             bool scan8 = (thrInBlock < CR_BIN_SQR / 32);
-            U32 scan8Mask = __ballot_sync(~0u, scan8);
+            U32 scan8Mask = __ballot_sync((unsigned long long)(~0u), scan8);
             if (scan8)
             {
                 int sum = s_tileEmitPrefixSum[(thrInBlock << 5) + 32];
@@ -654,7 +654,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
         {
             int tileInBin = tileInBin_base + thrInBlock;
             bool act = (tileInBin < CR_BIN_SQR);
-            U32 actMask = __ballot_sync(~0u, act);
+            U32 actMask = __ballot_sync((unsigned long long)(~0u), act);
             if (act)
             {
                 int tileX = tileInBin & (CR_BIN_SIZE - 1);
@@ -676,7 +676,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
                 if (segCount != 0)
                     tileSegCount[segIdx] = segCount;
 
-                U32 res = __ballot_sync(actMask, ofs >= 0 | force);
+                U32 res = __ballot_sync((unsigned long long)(actMask), ofs >= 0 | force);
                 if (threadIdx.x == 0)
                     s_scanTemp[0][(tileInBin >> 5) + 16] = __popc(res);
             }
@@ -688,7 +688,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
         __syncthreads();
 
         bool scan8 = (thrInBlock < CR_BIN_SQR / 32);
-        U32 scan8Mask = __ballot_sync(~0u, scan8);
+        U32 scan8Mask = __ballot_sync((unsigned long long)(~0u), scan8);
         if (scan8)
         {
             volatile U32* v = &s_scanTemp[0][thrInBlock + 16];
@@ -715,7 +715,7 @@ __device__ __inline__ void coarseRasterImpl(const CRParams p)
         {
             int tileInBin = tileInBin_base + thrInBlock;
             bool act = (tileInBin < CR_BIN_SQR) && (s_tileStreamCurrOfs[tileInBin] >= 0);
-            U32 actMask = __ballot_sync(~0u, act);
+            U32 actMask = __ballot_sync((unsigned long long)(~0u), act);
             if (act)
             {
                 int activeIdx = s_firstActiveIdx;
